fix: 消除Kotlin编译警告（废弃覆盖/冗余转换/不必要安全调用）
- 为废弃API覆盖添加 @Deprecated 注解：onUse、onStateReplaced、onStacksDropped、
  hasComparatorOutput、getComparatorOutput、getBackground、getCullingShape
- 移除冗余 Elvis 运算符（ModMachineRecipes.recipeType 返回非空）
- 移除不必要类型转换 as?、.toInt()、.toLong()、.toDouble()
- ContainerItemContext.withInitial → withConstant
- World.isChunkLoaded(BlockPos) → isChunkLoaded(x, z)
- SpriteAtlasTexture.BLOCK_ATLAS_TEXTURE → 显式 Identifier
- 移除不必要 !! 和安全调用（?.）
- registryEntry → Registries.ITEM.getId()
- 对废弃父类调用添加 @Suppress("DEPRECATION")

+++ core/src/client/kotlin/ic2_120/client/screen/ChunkLoaderScreen.kt
@@ -134,7 +134,7 @@ class ChunkLoaderScreen(
         }
 
         // 网格细线（半透明，仅区块边界）
-        val gridColor = 0x33FFFFFF.toInt()
+        val gridColor = 0x33FFFFFF
         for (i in 1 until GRID) {
             val lineX = mapLeft + i * CELL_SIZE
             context.fill(lineX, mapTop, lineX + 1, mapTop + MAP_SIZE, gridColor)
